Refactor AppStyle extensions
